cloud/relay/pyproject.toml
[project]
name = "lem-relay"
version = "0.1.0"
description = "Lem cloud relay server for WebSocket frame forwarding"
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "mypy>=1.13.0",
    "ruff>=0.7.0",
    "types-python-jose>=3.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
